requirements.txt
streamlit
langchain
langchain-openai
openai
langchain-core
langchain-chroma
python-dotenv
dotenv
chromadb
pydantic
pyyaml
